+++ changelog.md
@@ -17,6 +17,7 @@ Codion Change Log
 - NumberDocument, translate single minus sign to -1, in order to prevent minus sign input when negative numbers are not valid.
 - ColumnConditionPanel, now displays the selected operator in a tooltip when changed via the keyboard, if the view is SIMPLE and the operator combo box hidden.
 - FilterTable.Export moved to FilterTableModel, related refactoring.
+- NumberField.Builder.maximumFractionDigits() renamed fractionDigits().
 ### is.codion.framework.domain
 - KeyGenerator.Identity removed.
 - AttributeDefinition.Builder.valueRange(), minimumValue() and maximumValue() renamed range(), minimum() and maximum().